// src/app/interfaces/ui-config/models/movies.interface.ts
export interface MovieData {
  page: number;
  results: MovieResult[];
  total_pages: number;
  tottal_results: number;
}

export interface MovieResult {
  adult: boolean;
  backdrop_path: string;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  video: string;
  vote_average: number;
  vote_count: number;
}